Add MIT License to the project
Introduced an MIT License to the project, granting permissions for use, copy, modification, merging, publishing, distribution, sublicensing, and selling of the software. The license includes a 2024 copyright notice attributed to Hootan Hemmati and specifies that the software is provided "as is" without any warranty, limiting the liability of the authors.